% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormind-package.R
\docType{package}
\name{colormind-package}
\alias{colormind}
\alias{colormind-package}
\title{colormind: Generate Color Palettes With AI}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Use colormind.io API to generate aesthetically
    pleasing color palettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dmi3kno/colormind}
  \item Report bugs at \url{https://github.com/dmi3kno/colormind/issues}
}

}
\author{
\strong{Maintainer}: Dmytro Perepolkin \email{dperepolkin@gmail.com} (\href{https://orcid.org/0000-0001-8558-6183}{ORCID})

}
\keyword{internal}
