% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_colormind_colors}
\alias{get_colormind_colors}
\title{Retrieve color palette from colormind, give suggestions}
\usage{
get_colormind_colors(x = NULL, model = "default", verbose = TRUE)
}
\arguments{
\item{x}{character vector of valid colors in hex or string (color names) format,
which should serve as starting values for generated palette. Can contain missing values.
Maximum length of vector is 5.}

\item{model}{one of color models returned by `get_colormind_models()`. Default value is "default".}

\item{verbose}{optional argument for suppressing messages. Default is `TRUE``}
}
\value{
vector of colors in hex format
}
\description{
Retrieve color palette from colormind, give suggestions
}
\examples{
get_colormind_colors()

}
