% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_colormind_models}
\alias{get_colormind_models}
\title{Get currently available models}
\usage{
get_colormind_models()
}
\value{
character vector of color models
}
\description{
Get currently available models
}
\examples{
get_colormind_models()

}
