% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex2rgba}
\alias{hex2rgba}
\title{INTERNAL function for converting color to rgba}
\usage{
hex2rgba(x, alpha = TRUE)
}
\arguments{
\item{x}{vector of colors (named or hex values)}

\item{alpha}{logical whether to include transparency. Default is TRUE}
}
\description{
INTERNAL function for converting color to rgba
}
