% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_colormind_query}
\alias{prepare_colormind_query}
\title{INTERNAL function for making json string for colormind query}
\usage{
prepare_colormind_query(x, model)
}
\arguments{
\item{x}{vector of colors with missing values}

\item{model}{one of valid models for colormind}
}
\value{
properly formatted query string
}
\description{
INTERNAL function for making json string for colormind query
}
