% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rgba2hex}
\alias{rgba2hex}
\title{INTERNAL function for converting rgba to hex}
\usage{
rgba2hex(r, g, b, a)
}
\arguments{
\item{r, g, b, a}{vectors for r, g, b, a or a matrix of colors}
}
\description{
INTERNAL function for converting rgba to hex
}
