% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_colormind_models}
\alias{get_colormind_models}
\title{Get currently available models}
\usage{
get_colormind_models(
  user_agent = "colormind R package github.com/dmi3kno/colormind",
  ...
)
}
\arguments{
\item{user_agent}{Optional user agent string to be used when querying colormind.io}

\item{...}{other parameters passed to `config` argument of `httr::GET`}
}
\value{
character vector of color models
}
\description{
Get currently available models
}
\examples{
if(curl::has_internet())
  get_colormind_models()

}
